// src/app/core/models/candidate.model.ts
import {CatalogItemModel} from "./catalog.item.model";
import {UserModel} from "./user.model";
import {CandidateCertificateModel} from "./candidate.certificate.model";
import {CandidateDomainModel} from "./candidate.domain.model";
import {CandidateContactHistoryModel} from "./candidate.contact.history.model";
import {CandidateFileModel} from "./candidate.file.model";
import {CandidateContactModel} from "./candidate.contact.model";
import {CandidateSchoolModel} from "./candidate.school.model";

export class CandidateModel {
    candidate_id: number;
    candidate_code: string;
    full_name: string;
    birth_date: string;
    phone: string;
    email: string;
    address: string;
    gender: string;
    source_id: number;
    receive_time: string;
    jd_link: string;
    apply_position_id: number;
    apply_position: CatalogItemModel;
    technology: CatalogItemModel;
    status: CatalogItemModel;
    apply_position_name: string;
    level_id: number;
    domain_id: number;
    old_company: string;
    start_work_time: string;
    technology_id: number;
    technology_name: string;
    start_technology_time: string;
    max_literacy_id: number;
    literacy_english_id: number;
    in_charge_user_id: number;
    status_id: number;
    status_name: string;
    create_user_id: number;
    last_update_user_id: number;
    hr: string;
    create_date: Date;
    create_time: Date;
    createdAt: Date;
    checked: boolean = false;
    create_user: UserModel;
    in_charge_user: UserModel;
    certificates: CandidateCertificateModel[];
    domains: CandidateDomainModel[];
    contact_histories: CandidateContactHistoryModel[];
    files: CandidateFileModel[];
    contacts: CandidateContactModel[];
    schools: CandidateSchoolModel[];
    display: string;
    can_view_detail: boolean;
    nationality: string;
}